Split VRT MAX provider into __init__/constants/provider modules
Follow the recommended provider layout: __init__.py holds only setup(), the
browse/tile/tracklist constants move to constants.py, and the provider class plus
its id helpers move to provider.py (the GraphQL client stays in helpers.py). The
provider now owns SUPPORTED_FEATURES via its own __init__, matching other split
providers. Pure refactor: no behaviour change, tests updated to the new imports.
